Core/Qt: should fix bug 1138 (SR paths on Windows)
POSIX uses "/" as path separator while Windows uses "\". Search results
are sent in the Windows format. Some clients send search results as
simple file names without directories (or they have files in the root of
the share), which is OK but it led to this strange bug because
SearchResult::getFileName() actually intends to return the file or
directory base name, not the whole path. On POSIX it worked, because it
was overlooked that Util::getFileName() uses the native path separator
"/" which is illegal in Windows paths therefore never finding a
separator and always returning the whole path (as needed by
SearchResult::getFileName). On Windows however, Util::getFileName()
works as originally intended, which in turn breaks
SearchResult::getFileName() when it tries to split the path a second
time.

*phew* ('^_^)

Note: GUIs untested, please review