Revert changes from c7536437a452c49766bc4b4aff282acdd7466048
and make real fix. Quotation:
The p, q and g values can be set by calling DH_set0_pqg() and passing
the new values for p, q and g as parameters to the function. Calling
this function transfers the memory management of the values to the DH
object, and therefore the values that have been passed in should not
be freed directly after this function has been called. The q parameter
may be NULL.

https://www.openssl.org/docs/man1.1.0/crypto/DH_get0_pqg.html